+++ src/pytest_bdd_report/components/scenario.py
@@ -10,7 +10,7 @@ class Scenario:
     description: str
     tags: []
     steps: list[Step]
-    duration: int = 0
+    duration: float = 0.0
 
     def calculate_duration(self) -> None:
         duration = 0
